Fix captains algo
